harness: extract native_tools artifacts module
Move create/read/update/get/grant_artifact + helpers
(artifact_content_bytes, resolve_artifact_uri, artifact_json, etc.)
into native_tools/artifacts.rs. Re-export for a2a.rs/tools compat.
mod.rs 4251 -> 3834 LoC.
